remove all drone code and jtech references
