+++ README.md
@@ -1,62 +1,47 @@
-# Drone Project
+# Streamerz
+
+A playground of video processing examples in Akka streams and Scala (with some Kafka).
+
+It's always been said that the best way to learn a topic through meme engineering.
+This set of libraries enables all the fun you never wanted to have.
 
 
-## How to start drone feed 
 
-### Prerequisites:
+## Examples
 
+1.  Rendering your webcam to the terminal in Ascii art
 
-```
-brew install kafka 
-brew install zookeper 
-```
+    ```
+    $ sbt "examples/runMain examples.AsciiWebcam"
+    ```
 
+2.  Render video (and audio) to the terminal in Ascii Art
 
-### Steps to run: 
+    ```
+    $ sbt "examples/runMain examples.AsciiVideo"
+    ```
 
-1.  Start zookeeper:
+3. Create a swing video player that can play/pause an mp4
+
+    ```
+    $ sbt "examples/runMain examples.VideoPlayer"
+    ```
+
+4. Run a web server that streams your webcam to web browsers in Ascii art
+
+    ```
+    $ brew install kafka 
+    $ brew install zookeper
+    ```
 
     ```
     $ zkServer start
+    $ kafka-server-start.sh /usr/local/etc/kafka/server.properties &
+    $ sbt "examples/runMain examples.asciiweb.feed.AsciiImageProducer" &
+    $ sbt "examples/runMain examples.asciiweb.ws.Main" &
     ```
-
-2. Run Kafka broker:
-
-   ```
-   $ kafka-server-start.sh /usr/local/etc/kafka/server.properties
-   ```
-
-   To monitor traffic, run Kafka console consumer as follows (in a different shell): 
-   
-   ```
-   $ kafka-console-consumer.sh —-topic DRONE_FEED —-zookeeper localhost:2181
-   ```
-   
-3. To use the webcam as a source: 
-   - Run:
-  
-     ```
-     $ sbt "examples/runMain org.jtech.drone.feed.AsciiImageProducer"
-     ```
-    
-   To use the drone as a source:
-    
-   - Turn on the drone
-   - After its initial checks, connect to its WiFi hotspot
-   - Run:
-   
-     ```
-     $ sbt "examples/runMain org.jtech.drone.feed.AsciiImageProducer drone"
-     ```
-     
-4. Launch the web server:
-
-   ```
-   $ sbt "examples/runMain org.jtech.drone.ws.Main"
-   ```
-
-You can now check your webcam/drone feed as an Ascii-art animation on `http://localhost:5000`
-
+ 
+  You can now check your webcam/drone feed as an Ascii-art animation on `http://localhost:8080`
 
 ## License
