+++ examples/src/main/resources/application.conf
@@ -8,7 +8,7 @@ akka {
 
 ordina.jtech.drone {
   ws {
-    port = 5000
+    port = 8080
   }
 
   kafka {
@@ -17,7 +17,7 @@ ordina.jtech.drone {
     zookeeper.hosts = "localhost:2181"
 
     consumer {
-      group.id = "org.jtech.drone.feed.consumer"
+      group.id = "examples.asciiweb.feed.consumer"
     }
   }
 }
