--- examples/src/main/resources/web/ws-ping.html
@@ -1,24 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <title>Websocket Ping Stream</title>
-</head>
-<body>
-
-<script src="jquery-1.10.2.js"></script>
-<script>
-    var asciiData = new WebSocket("ws://localhost:5000/ping");
-    asciiData.onopen = function(event) {
-        console.log("Connected!", event);
-    };
-    asciiData.onmessage = function(event) {
-        var data = event.data;
-        console.log("Got following data", data);
-        $("body").html(data);
-    };
-
-
-</script>
-
-</body>
-</html>
