+++ examples/src/main/scala/examples/WebSocketAsciiVideo.scala
